#1 : Create a connector for renewing Apache HTTP Basic passwords
* Add checks in the scripts

+++ scripts/rollback_htdigest.sh
@@ -1,7 +1,14 @@
 #!/bin/sh
 
+# exit when any command fails
+set -e
+
 HTDIGEST_PATH=/etc/apache2/htdigest
 HTDIGEST_BACKUP_PATH=/etc/apache2/htdigest.old
 
 # Rollback the file
-mv $HTDIGEST_BACKUP_PATH $HTDIGEST_PATH
+if [ -f $HTDIGEST_BACKUP_PATH ] ; then
+  mv $HTDIGEST_BACKUP_PATH $HTDIGEST_PATH
+else
+  echo 'No backup file found.' >&2
+fi
