+++ scripts/update_htdigest.sh
@@ -1,5 +1,8 @@
 #!/bin/sh
 
+# exit when any command fails
+set -e
+
 # Verify that we have enough arguments
 if [ "$#" -ge 3 ] ; then
 
@@ -14,11 +17,14 @@ HTDIGEST_PATH=/etc/apache2/htdigest
 HTDIGEST_BACKUP_PATH=/etc/apache2/htdigest.old
 
 # First, backup the htdigest file
-cp $HTDIGEST_PATH $HTDIGEST_BACKUP_PATH
+if [ -f $HTDIGEST_PATH ] ; then
+  cp $HTDIGEST_PATH $HTDIGEST_BACKUP_PATH
+fi
 
 (echo -n "$HTDIGEST_USER:$HTDIGEST_REALM:" \
- && echo -n "$HTDIGEST_USER:$HTDIGEST_REALM:$HTDIGEST_PASSWORD" | $MD5SUM | $AWK '{print $1}') > htdigest
+ && echo -n "$HTDIGEST_USER:$HTDIGEST_REALM:$HTDIGEST_PASSWORD" \
+ | $MD5SUM | $AWK '{print $1}') > $HTDIGEST_PATH
 else
-  echo "Not enough arguments" >&2
-  echo "Usage : update_htdigest.sh user realm password"
+  echo 'Not enough arguments' >&2
+  echo 'Usage : update_htdigest.sh user realm password'
 fi
